clearway-for-telegram/pyproject.toml
[project]
name = "clearway-for-telegram"
version = "0.1.0"
description = "Clearway bridge for Telegram using Telethon"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "telethon>=1.35.0",
    "python-dotenv>=1.0.1",
]
